Renamed `ExcludeProps` to `IgnoredProperties` in `AutoDtoAttribute` classes
Renamed the property `ExcludeProps` to `IgnoredProperties` in both `AutoDtoAttribute` and `AutoDtoAttribute<T>` classes. This change is also reflected in the constructors of these classes, where the parameter `excludeProps` has been renamed to `ignoredProperties`. Updated the `Gen-Dto.md` file to reflect these changes in the usage instructions for the `AutoDto` and `AutoDto<T>` attributes.